requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-email-users==1.1.0
django-oauth-toolkit==1.5.0
django-oidc-provider==0.7.0
djangorestframework==3.12.4
djangorestframework-oauth==1.1.0
docopt==0.6.2
future==0.18.2
hashids==1.3.1
httpie==2.4.0
idna==2.10
jwcrypto==0.8
networkx==2.5.1
oauthlib==3.1.0
pedigree==1.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.9.0
pyjwkest==1.4.2
PySocks==1.7.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
views==0.3
